% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_protocol.R
\name{analyze_stemp}
\alias{analyze_stemp}
\title{Analyze STeMP output}
\usage{
analyze_stemp(file)
}
\arguments{
\item{file}{Path to the downloaded .csv from the STeMP WebAPP.}
}
\value{
Table listing possible problems identified from the filled protocol.
}
\description{
Analyze STeMP output
}
