% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_geodist_classification}
\alias{calculate_geodist_classification}
\title{Calculate Geodistance Classification}
\usage{
calculate_geodist_classification(samples_sf, area_sf)
}
\arguments{
\item{samples_sf}{sf object of sample locations.}

\item{area_sf}{sf object of spatial area.}
}
\value{
Character classification "random" or "clustered".
}
\description{
Calculate Geodistance Classification
}
