% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{firstup}
\alias{firstup}
\title{Capitalize First Letter of a String}
\usage{
firstup(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Character vector with first letter capitalized.
}
\description{
Converts the first character of a string to uppercase.
}
