% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geo_map}
\alias{geo_map}
\title{Render a Geo Map Plot to Output}
\usage{
geo_map(
  output,
  element_id,
  geo_metadata = NULL,
  what = c("samples_sf", "training_area_sf", "prediction_area_sf")
)
}
\arguments{
\item{output}{Shiny output object.}

\item{element_id}{Output ID for the plot.}

\item{geo_metadata}{Reactive list containing spatial data.}

\item{what}{Character specifying which spatial data to use.}
}
\description{
Render a Geo Map Plot to Output
}
