% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geodist_plot}
\alias{geodist_plot}
\title{Render Geodistance Plot to Output}
\usage{
geodist_plot(
  output,
  element_id,
  geo_metadata = NULL,
  objective = c("Model and prediction", "Model only")
)
}
\arguments{
\item{output}{Shiny output object.}

\item{element_id}{Output ID for the plot.}

\item{geo_metadata}{Reactive list containing spatial data.}

\item{objective}{Character indicating which area to use.}
}
\description{
Render Geodistance Plot to Output
}
