% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_value}
\alias{get_value}
\title{Get Value from Uploaded Input or Fallback}
\usage{
get_value(uploaded_value, fallback_fn)
}
\arguments{
\item{uploaded_value}{Value from uploaded input.}

\item{fallback_fn}{Function to call if uploaded_value is NULL.}
}
\value{
The uploaded value, fallback function result, or NULL.
}
\description{
Get Value from Uploaded Input or Fallback
}
