% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inputWithHoverInfo}
\alias{inputWithHoverInfo}
\title{Add Tooltip Info to Shiny Input Tag}
\usage{
inputWithHoverInfo(inputTag, info_text)
}
\arguments{
\item{inputTag}{A Shiny input tag.}

\item{info_text}{Text to show in tooltip.}
}
\value{
Modified input tag with tooltip added.
}
\description{
Add Tooltip Info to Shiny Input Tag
}
