% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_protocol.R
\name{protocol_analyze}
\alias{protocol_analyze}
\title{Analyze STeMP output}
\usage{
protocol_analyze(protocol, render = TRUE)
}
\arguments{
\item{protocol}{A data.frame containing the downloaded STeMP object.}

\item{render}{Should the resulting table be rendered? By default TRUE.}
}
\value{
A list containing the warnings as plain text, as well as the rendered table if `render` is TRUE.
}
\description{
Analyze STeMP output
}
