% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_read.R
\name{protocol_read}
\alias{protocol_read}
\title{Read a protocol object}
\usage{
protocol_read(protocol_path)
}
\arguments{
\item{protocol_path}{Path to the protocol file. Currently only CSV files are supported.}
}
\value{
A data frame of the protocol which can be used by [protocol_analyze()].
}
\description{
Read a protocol object
}
