% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_figure}
\alias{save_figure}
\title{Save ggplot Figure to File}
\usage{
save_figure(figure, element_id)
}
\arguments{
\item{figure}{ggplot object.}

\item{element_id}{Character ID to determine filename.}
}
\description{
Save ggplot Figure to File
}
