% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unwrap_metadata}
\alias{unwrap_metadata}
\title{Unwrap Reactive Values in Metadata List}
\usage{
unwrap_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{List containing reactive or non-reactive elements.}
}
\value{
List with all elements unwrapped (reactive evaluated).
}
\description{
Unwrap Reactive Values in Metadata List
}
