% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Add Tooltip to Shiny Input if Info Text Provided}
\usage{
with_tooltip(inputTag, info_text = NULL)
}
\arguments{
\item{inputTag}{A Shiny input tag.}

\item{info_text}{Optional tooltip text.}
}
\value{
Shiny input tag with tooltip if info_text is not NULL/empty.
}
\description{
Add Tooltip to Shiny Input if Info Text Provided
}
